requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.2.1
django-browser-reload==1.18.0
django-tinymce==4.1.0
django-unfold==0.65.0
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
webencodings==0.5.1
whitenoise==6.9.0
